requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
azure-core==1.29.4
azure-storage-blob==12.18.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
Django
django-livereload-server==0.4
django-livesync==0.5
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
flatbuffers==23.5.26
fontawesomefree==6.4.2
fonttools==4.42.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
ml-dtypes==0.2.0
mpld3==0.5.9
mysql-connector-python==8.1.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
phonenumbers==8.13.17
Pillow==10.0.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydicom==2.4.3
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
tk==0.1.0
tornado==6.3.3
typing_extensions==4.8.0
urllib3==1.26.16
watchdog==3.0.0
websocket-client==1.6.1
Werkzeug==2.3.7
whitenoise==6.5.0
wrapt==1.14.1